+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added `bit_count()` function, pr #405.
 * Allow `return;` statement without arguments _(implicitly return nil)_, pr #404.
+* Function `nse();` should not error when a side-effect outside nse() is enforced
 
 # v1.7.3